plit out moduletool testcases
