+++ tests/moduletool/test_basics.py
@@ -1,0 +1,86 @@
+"""
+    Copyright 2017 Inmanta
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+    Contact: code@inmanta.com
+"""
+import re
+
+from pkg_resources import parse_version
+
+from inmanta import module
+from inmanta.moduletool import ModuleTool
+
+
+def test_versioning():
+    mt = ModuleTool()
+
+    newversion = mt.determine_new_version(parse_version("1.2.3"), None, False, False, True, False)
+    assert str(newversion) == "1.2.4"
+    newversion = mt.determine_new_version(parse_version("1.2.3"), None, False, True, False, False)
+    assert str(newversion) == "1.3.0"
+    newversion = mt.determine_new_version(parse_version("1.2.3"), None, True, False, False, False)
+    assert str(newversion) == "2.0.0"
+    newversion = mt.determine_new_version(parse_version("1.2.3"), None, True, True, False, False)
+    assert newversion is None
+    newversion = mt.determine_new_version(parse_version("1.2.3"), None, True, False, True, False)
+    assert newversion is None
+    newversion = mt.determine_new_version(parse_version("1.2.3"), None, True, True, True, False)
+    assert newversion is None
+    newversion = mt.determine_new_version(parse_version("1.2.3.dev025"), None, False, False, True, False)
+    assert str(newversion) == "1.2.3"
+    newversion = mt.determine_new_version(parse_version("1.2.3.dev025"), None, False, False, False, False)
+    assert str(newversion) == "1.2.3"
+
+    newversion = mt.determine_new_version(parse_version("1.2.3"), None, False, False, True, True)
+    assert re.search("1.2.4.dev[0-9]+", str(newversion))
+    newversion = mt.determine_new_version(parse_version("1.2.3"), None, False, True, False, True)
+    assert re.search("1.3.0.dev[0-9]+", str(newversion))
+    newversion = mt.determine_new_version(parse_version("1.2.3"), None, True, False, False, True)
+    assert re.search("2.0.0.dev[0-9]+", str(newversion))
+    newversion = mt.determine_new_version(parse_version("1.2.3"), None, True, True, False, True)
+    assert newversion is None
+    newversion = mt.determine_new_version(parse_version("1.2.3"), None, True, False, True, True)
+    assert newversion is None
+    newversion = mt.determine_new_version(parse_version("1.2.3"), None, True, True, True, True)
+    assert newversion is None
+    newversion = mt.determine_new_version(parse_version("1.2.3.dev025"), None, False, False, True, True)
+    assert re.search("1.2.3.dev[0-9]+", str(newversion))
+    newversion = mt.determine_new_version(parse_version("1.2.3.dev025"), None, False, False, False, True)
+    assert re.search("1.2.3.dev[0-9]+", str(newversion))
+
+
+def test_rewrite(tmpdir):
+    module_path = tmpdir.join("mod").mkdir()
+    model = module_path.join("model").mkdir()
+    model.join("_init.cf").write("\n")
+
+    module_yml = module_path.join("module.yml")
+    module_yml.write(
+        """
+name: mod
+license: ASL
+version: 1.2
+compiler_version: 2017.2
+    """
+    )
+
+    mod = module.Module(None, module_path.strpath)
+
+    assert mod.version == "1.2"
+    assert mod.compiler_version == "2017.2"
+
+    mod.rewrite_version("1.3.1")
+    assert mod.version == "1.3.1"
+    assert mod.compiler_version == "2017.2"
